pulled state out with useContext hook and restyled components
